feature: setup basic nav style structure
